Voltando a ordem do VND distribuído para: r = 1: positiveCliqueMove r = 2: moveCluster1opt r = 3: swapCluster1opt r = 4: twoMoveCluster r = 5: moveVertex1opt. Incluindo o parametro percentual POS_CLIQUE_PERC_RELAX para relaxação das cliques geradas pela vizinhança positiveCliqueMove.

+++ grasp/src/resolution/ils/splitgraph/include/ImbalanceSubgraphParallelILS.h
@@ -226,7 +226,7 @@ protected:
 	std::list<VertexDegree> calculateDegreesInsideCluster(SignedGraph *g,
 			Clustering& globalClustering, long clusterX);
 
-	long chooseRandomVertex(std::list<VertexDegree>& vertexList, int x);
+	long chooseRandomVertex(std::list<VertexDegree>& vertexList, long x);
 
 };
